Moved equip attachment command
